[Automation] Date Type Monitor (#10)
Leveraging the WindowedQueryMonitor we enable the possibility to define a TimeWindow for an entire day and make configuring a "yesterday" time window very easy. This will be the configuration type for our Unusual Solver Slippage query because its the easiest to run as a cronjob (i.e. once per day, it will reveal any alerts regarding slippage for the previous day).